MA-18051 avoid overflow of in partition size calculation
If a partition is not less than 4GB, to avoid the overflow issue when
calculate the partition size in bytes, change the value of partition
length in block size to the type of "unsigned long".

(cherry picked from commit c4e9be08d90ca59531542c97bd8fe16eab2ab099)
(cherry picked from commit 6c1471828256f5c95d3043099fcc00906726fdd3)

+++ include/fb_fsl.h
@@ -167,7 +167,7 @@ struct fastboot_ptentry {
 	/* The start wrt the nand part, must be multiple of nand block size */
 	unsigned int start;
 	/* The length of the partition, must be multiple of nand block size */
-	unsigned int length;
+	unsigned long length;
 	/* Controls the details of how operations are done on the partition
 	   See the FASTBOOT_PTENTRY_FLAGS_*'s defined below */
 	unsigned int flags;
